Fix hue calculation (#4)
* Fixed hue calculation and improved tests around that logic.

* Change amplitude to float (as that is the type we primarily deal with)

* Omit frequencies above 120hz from the hue.

+++ src/main/kotlin/sound/frequencyBins/FrequencyBinsFactory.kt
@@ -44,7 +44,7 @@ class FrequencyBinsFactory(
     private fun getFrequencyBin(index: Int, amplitude: Double, sampleSize: Int, sampleRate: Float): FrequencyBin {
         return FrequencyBin(
             frequency = getFrequency(index, sampleSize, sampleRate),
-            amplitude = amplitude
+            amplitude = amplitude.toFloat()
         )
     }
